+++ frontend/src/metabase/query_builder/components/view/sidebars/SummarizeSidebar/SummarizeSidebar.tsx
@@ -1,20 +1,11 @@
-import { useCallback, useMemo, useState } from "react";
+import { useCallback } from "react";
 import { t } from "ttag";
 
 import { color } from "metabase/lib/colors";
-import * as Lib from "metabase-lib";
+import type * as Lib from "metabase-lib";
 
-import { AddAggregationButton } from "./AddAggregationButton";
-import { AggregationItem } from "./AggregationItem";
-import { BreakoutColumnList } from "./BreakoutColumnList";
-import {
-  AggregationsContainer,
-  ColumnListContainer,
-  SectionTitle,
-  SidebarView,
-} from "./SummarizeSidebar.styled";
-
-const STAGE_INDEX = -1;
+import { SummarizeContent, useSummarizeQuery } from "./SummarizeContent";
+import { SidebarView } from "./SummarizeSidebar.styled";
 
 interface SummarizeSidebarProps {
   className?: string;
@@ -29,86 +20,18 @@ export function SummarizeSidebar({
   onQueryChange,
   onClose,
 }: SummarizeSidebarProps) {
-  const [isDefaultAggregationRemoved, setDefaultAggregationRemoved] =
-    useState(false);
-
-  const query = useMemo(
-    () => getQuery(initialQuery, isDefaultAggregationRemoved),
-    [initialQuery, isDefaultAggregationRemoved],
-  );
-
-  const aggregations = Lib.aggregations(query, STAGE_INDEX);
-  const hasAggregations = aggregations.length > 0;
-
-  const handleAddAggregation = useCallback(
-    (aggregation: Lib.Aggregable) => {
-      const nextQuery = Lib.aggregate(query, STAGE_INDEX, aggregation);
-      onQueryChange(nextQuery);
-    },
-    [query, onQueryChange],
-  );
-
-  const handleUpdateAggregation = useCallback(
-    (aggregation: Lib.AggregationClause, nextAggregation: Lib.Aggregable) => {
-      const nextQuery = Lib.replaceClause(
-        query,
-        STAGE_INDEX,
-        aggregation,
-        nextAggregation,
-      );
-      onQueryChange(nextQuery);
-    },
-    [query, onQueryChange],
-  );
-
-  const handleRemoveAggregation = useCallback(
-    (aggregation: Lib.AggregationClause) => {
-      const nextQuery = Lib.removeClause(query, STAGE_INDEX, aggregation);
-      const nextAggregations = Lib.aggregations(nextQuery, STAGE_INDEX);
-      if (nextAggregations.length === 0) {
-        setDefaultAggregationRemoved(true);
-      }
-      onQueryChange(nextQuery);
-    },
-    [query, onQueryChange],
-  );
-
-  const handleAddBreakout = useCallback(
-    (column: Lib.ColumnMetadata) => {
-      const nextQuery = Lib.breakout(query, STAGE_INDEX, column);
-      onQueryChange(nextQuery);
-    },
-    [query, onQueryChange],
-  );
-
-  const handleUpdateBreakout = useCallback(
-    (clause: Lib.BreakoutClause, column: Lib.ColumnMetadata) => {
-      const nextQuery = Lib.replaceClause(query, STAGE_INDEX, clause, column);
-      onQueryChange(nextQuery);
-    },
-    [query, onQueryChange],
-  );
-
-  const handleRemoveBreakout = useCallback(
-    (column: Lib.ColumnMetadata) => {
-      const { breakoutPosition } = Lib.displayInfo(query, STAGE_INDEX, column);
-      if (typeof breakoutPosition === "number") {
-        const breakouts = Lib.breakouts(query, STAGE_INDEX);
-        const clause = breakouts[breakoutPosition];
-        const nextQuery = Lib.removeClause(query, STAGE_INDEX, clause);
-        onQueryChange(nextQuery);
-      }
-    },
-    [query, onQueryChange],
-  );
-
-  const handleReplaceBreakouts = useCallback(
-    (column: Lib.ColumnMetadata) => {
-      const nextQuery = Lib.replaceBreakouts(query, STAGE_INDEX, column);
-      onQueryChange(nextQuery);
-    },
-    [query, onQueryChange],
-  );
+  const {
+    query,
+    aggregations,
+    hasAggregations,
+    handleAddAggregation,
+    handleUpdateAggregation,
+    handleRemoveAggregation,
+    handleAddBreakout,
+    handleUpdateBreakout,
+    handleRemoveBreakout,
+    handleReplaceBreakouts,
+  } = useSummarizeQuery(initialQuery, onQueryChange);
 
   const handleDoneClick = useCallback(() => {
     onQueryChange(query);
@@ -122,49 +45,18 @@ export function SummarizeSidebar({
       color={color("summarize")}
       onDone={handleDoneClick}
     >
-      <AggregationsContainer>
-        {aggregations.map(aggregation => (
-          <AggregationItem
-            key={
-              Lib.displayInfo(query, STAGE_INDEX, aggregation).longDisplayName
-            }
-            query={query}
-            aggregation={aggregation}
-            onUpdate={nextAggregation =>
-              handleUpdateAggregation(aggregation, nextAggregation)
-            }
-            onRemove={() => handleRemoveAggregation(aggregation)}
-          />
-        ))}
-        <AddAggregationButton
-          query={query}
-          onAddAggregation={handleAddAggregation}
-        />
-      </AggregationsContainer>
-      {hasAggregations && (
-        <ColumnListContainer>
-          <SectionTitle>{t`Group by`}</SectionTitle>
-          <BreakoutColumnList
-            query={query}
-            onAddBreakout={handleAddBreakout}
-            onUpdateBreakout={handleUpdateBreakout}
-            onRemoveBreakout={handleRemoveBreakout}
-            onReplaceBreakout={handleReplaceBreakouts}
-          />
-        </ColumnListContainer>
-      )}
+      <SummarizeContent
+        query={query}
+        aggregations={aggregations}
+        hasAggregations={hasAggregations}
+        onAddAggregation={handleAddAggregation}
+        onUpdateAggregation={handleUpdateAggregation}
+        onRemoveAggregation={handleRemoveAggregation}
+        onAddBreakout={handleAddBreakout}
+        onUpdateBreakout={handleUpdateBreakout}
+        onRemoveBreakout={handleRemoveBreakout}
+        onReplaceBreakouts={handleReplaceBreakouts}
+      />
     </SidebarView>
   );
-}
-
-function getQuery(query: Lib.Query, isDefaultAggregationRemoved: boolean) {
-  const hasAggregations = Lib.aggregations(query, STAGE_INDEX).length > 0;
-  const shouldAddDefaultAggregation =
-    !hasAggregations && !isDefaultAggregationRemoved;
-
-  if (!shouldAddDefaultAggregation) {
-    return query;
-  }
-
-  return Lib.aggregateByCount(query);
 }
